Rename folder and file to match original code
Rename "Project Saving" to "ProjectSaving" and rename
"jucer_ProjectExport_XCode.h" to "jucer_ProjectExport_Xcode.h" to match
the code of JUCE 5.2.1, which they are copied from.
