+++ cmake/XcassetsBuilder/main.cpp
@@ -1,4 +1,4 @@
-// Copyright (C) 2019  Alain Martin
+// Copyright (C) 2019-2020  Alain Martin
 //
 // This file is part of FRUT.
 //
@@ -17,7 +17,7 @@
 
 #include "Source/jucer_Headers.h"
 
-#include "Source/Project Saving/jucer_ProjectExporter.h"
+#include "Source/ProjectSaving/jucer_ProjectExporter.h"
 
 #include <iostream>
 #include <string>
